Connections are working
Ran some curl tests that post and get with firebase db

The firebase db I setup shows the info in it that was posted

Docker container builds and runs

After the container is running, the localhost:8080 link works for viewing the end points on the web

Havn't deployed to heroku just because it can get costly so we should wait to deploy until we have a mostly built api and are ready for working on frontend connections

The firebase-config.json file is something that I have and can distribute to everyone. I will want to look into this and see if it can be added in as a github secret or something like that so that it is there for everyone without actually being in the github repo.